Rename getUsers module to queries/users and update import paths
